% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkit.R
\docType{package}
\name{radminkit}
\alias{radminkit}
\title{radminkit: A Shiny dashboard template based on Adminkit}
\description{
A Shiny dashboard template based on Adminkit.
Uses the latest Bootstrap 5 release, which includes many new components.
}
