% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{radminkitBadge}
\alias{radminkitBadge}
\title{Create a badge}
\usage{
radminkitBadge(
  type = "secondary",
  intent = FALSE,
  pill = FALSE,
  darktext = FALSE,
  text = NULL
)
}
\arguments{
\item{type}{Type of badge. For included Bootstrap options, see details}

\item{intent}{(boolean) Is the badge meant to signal a message to the user?}

\item{pill}{(boolean) Should the badge have a rounded pill shape?}

\item{darktext}{(boolean) Should the badge use dark text? This is often
useful for brightly coloured badges.}

\item{text}{Text to display in the badge}
}
\description{
Create a badge
}
\details{
\subsection{Type}{

Valid options for type include:
\itemize{
\item primary
\item secondary
\item success
\item danger
\item warning
\item info
\item light
\item dark
}
}

\subsection{Intent}{

This flag is to indicate whether your badge is meant to alert the user.
For example, did you choose the danger badge for the colour scheme, or to
warn the user of something. If it is the latter, this flag will add an
accessibility class (\code{.visually-hidden}) to the badge.
}
}
\seealso{
https://getbootstrap.com/docs/5.0/badge/
}
