% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitBody.R
\name{radminkitBody}
\alias{radminkitBody}
\title{Create a radminkit body}
\usage{
radminkitBody(...)
}
\arguments{
\item{...}{Elements to put in the body of the dashboard}
}
\description{
Create a radminkit body
}
