% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{radminkitCard}
\alias{radminkitCard}
\title{Create a Bootstrap 5 card}
\usage{
radminkitCard(
  ...,
  title = NULL,
  subtitle = NULL,
  header = NULL,
  footer = NULL,
  text
)
}
\arguments{
\item{...}{Additional card parameters}

\item{title}{Title of the card.}

\item{subtitle}{(Optional) Card subtitle.}

\item{header}{(Optional) Card header.}

\item{footer}{(Optional) Card Footer}

\item{text}{Main body of the card.}
}
\description{
Create a Bootstrap 5 card
}
\seealso{
https://getbootstrap.com/docs/5.0/components/card/
}
