% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitPage.R
\name{radminkitPage}
\alias{radminkitPage}
\title{Create a radminkit page}
\usage{
radminkitPage(header, sidebar, body = NULL, title = NULL, loader = TRUE)
}
\arguments{
\item{header}{A header created with \code{\link{radminkitHeader}}}

\item{sidebar}{A sidebar created with \code{\link{radminkitSidebar}}}

\item{title}{Page title. This is the name that will appear in the browser tab}

\item{loader}{Loading indicator on top of page}
}
\description{
Create a radminkit page
}
