% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitSidebar - backup.R,
%   R/radminkitSidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Create a section entry}
\usage{
sidebarItem(
  ...,
  label = NULL,
  tabName = NULL,
  icon = NULL,
  itemTags = NULL,
  href = NULL,
  newTab = TRUE
)

sidebarItem(
  ...,
  label = NULL,
  tabName = NULL,
  icon = NULL,
  itemTags = NULL,
  href = NULL,
  newTab = TRUE
)
}
\arguments{
\item{tabName}{Name of the tab}

\item{icon}{(Optional) icon to display next to the tab name}

\item{itemTags}{(Optional) tags for the tab}

\item{href}{Link to tab}

\item{subitems}{Item(s) to add to the sidebar}
}
\description{
Create a section entry

Create a section entry
}
\details{
Radminkit uses \href{https://feathericons.com/}{Feather icons}; refer to their site for the full catalogue of options.
}
