% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitSidebar - backup.R,
%   R/radminkitSidebar.R
\name{sidebarSection}
\alias{sidebarSection}
\title{Create a sidebar section to split menus by category.}
\usage{
sidebarSection(label = "Section Label")

sidebarSection(label = "Section Label")
}
\arguments{
\item{label}{Section name}

\item{...}{Optional items to put in the sidebar. See Details sections}
}
\description{
Create a sidebar section to split menus by category.

Create a sidebar section to split tabs by category.
}
\details{
Possibilities include:
}
