% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitSidebar.R
\name{sidebarSubItem}
\alias{sidebarSubItem}
\title{Create section sub-entry}
\usage{
sidebarSubItem(tabName = NULL, href = NULL, icon = NULL)
}
\arguments{
\item{tabName}{Name of the subtab}

\item{href}{Link to tab}

\item{icon}{(Optional) icon to display next to the tab name}
}
\description{
Create section sub-entry
}
