% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitSidebar - backup.R,
%   R/radminkitSidebar.R
\name{sidebarTabs}
\alias{sidebarTabs}
\title{Create a sidebar tab menu}
\usage{
sidebarTabs(...)

sidebarTabs(...)
}
\arguments{
\item{...}{Items to add to the tablist}
}
\description{
Can add a section, menu, and submenu.

Can add a section, menu, and submenu.
}
