% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitHeader.R
\name{radminkitHeader}
\alias{radminkitHeader}
\title{Create a radminkit header}
\usage{
radminkitHeader(disable = FALSE, body = NULL, ...)
}
\arguments{
\item{disable}{Disable the header}

\item{body}{A \code{\link{radminkitBody}} item}

\item{...}{Additional components to insert in the header. See details section}
}
\description{
Create a radminkit header
}
\details{
Possibilities include:
A dropdownmenu
}
