% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitSidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Create a section entry}
\usage{
sidebarItem(
  ...,
  label = NULL,
  tabName = NULL,
  icon = NULL,
  itemTags = NULL,
  href = NULL,
  newTab = TRUE
)
}
\arguments{
\item{...}{(Optional) additional items to add to the tab
(e.g. a \code{\link{sidebarSubItem}})}

\item{label}{Text to display on the tab link}

\item{tabName}{Name of the tab}

\item{icon}{(Optional) icon to display next to the tab name}

\item{itemTags}{(Optional) tags for the tab}

\item{href}{Link to tab}

\item{newTab}{If \code{href} is supplied, should the link open in a new
browser tab?}
}
\description{
Create a section entry
}
\details{
Radminkit uses \href{https://feathericons.com/}{Feather icons};
refer to their site for the full catalogue of options.
}
