% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radminkitSidebar.R
\name{sidebarSection}
\alias{sidebarSection}
\title{Create a sidebar section to split tabs by category.}
\usage{
sidebarSection(label = "Section Label")
}
\arguments{
\item{label}{Section name}
}
\description{
Create a sidebar section to split tabs by category.
}
